test: refactor op tests
