+++ test/ops/hybrid/test_parametric_ops_single_qumode.py
@@ -9,6 +9,7 @@ import hybridlane as hqml
 @pytest.mark.unit
 class TestConditionalRotation:
     def test_init(self):
+        """Test the __init__ method."""
         op = hqml.ConditionalRotation(0.5, wires=[0, 1])
         assert op.name == "ConditionalRotation"
         assert op.num_params == 1
@@ -17,18 +18,21 @@ class TestConditionalRotation:
         assert op.wires == qml.wires.Wires([0, 1])
 
     def test_adjoint(self):
+        """Test the adjoint method."""
         op = hqml.ConditionalRotation(0.5, wires=[0, 1])
         adj_op = op.adjoint()
         assert isinstance(adj_op, hqml.ConditionalRotation)
         assert adj_op.parameters[0] == -0.5
 
     def test_pow(self):
+        """Test the pow method."""
         op = hqml.ConditionalRotation(0.5, wires=[0, 1])
         pow_op = op.pow(2)
         assert isinstance(pow_op[0], hqml.ConditionalRotation)
         assert pow_op[0].parameters[0] == 1.0
 
     def test_simplify(self):
+        """Test the simplify method."""
         op = hqml.ConditionalRotation(0, wires=[0, 1])
         simplified_op = op.simplify()
         assert isinstance(simplified_op, qml.Identity)
@@ -39,30 +43,9 @@ class TestConditionalRotation:
 
 
 @pytest.mark.unit
-class TestConditionalParity:
-    def test_init(self):
-        op = hqml.ConditionalParity(wires=[0, 1])
-        assert op.name == "ConditionalParity"
-        assert op.num_params == 0
-        assert op.num_wires == 2
-        assert op.parameters == []
-        assert op.wires == qml.wires.Wires([0, 1])
-
-    def test_decomposition(self):
-        op = hqml.ConditionalParity(wires=[0, 1])
-        decomp = op.decomposition()
-        assert len(decomp) == 1
-        assert isinstance(decomp[0], hqml.ConditionalRotation)
-
-    def test_adjoint(self):
-        op = hqml.ConditionalParity(wires=[0, 1])
-        adj_op = op.adjoint()
-        assert isinstance(adj_op, hqml.ConditionalRotation)
-
-
-@pytest.mark.unit
 class TestSelectiveQubitRotation:
     def test_init(self):
+        """Test the __init__ method."""
         op = hqml.SelectiveQubitRotation(0.5, 0.3, 1, wires=[0, 1])
         assert op.name == "SelectiveQubitRotation"
         assert op.num_params == 2
@@ -72,18 +55,21 @@ class TestSelectiveQubitRotation:
         assert op.wires == qml.wires.Wires([0, 1])
 
     def test_adjoint(self):
+        """Test the adjoint method."""
         op = hqml.SelectiveQubitRotation(0.5, 0.3, 1, wires=[0, 1])
         adj_op = op.adjoint()
         assert isinstance(adj_op, hqml.SelectiveQubitRotation)
         assert adj_op.parameters == [-0.5, 0.3]
 
     def test_pow(self):
+        """Test the pow method."""
         op = hqml.SelectiveQubitRotation(0.5, 0.3, 1, wires=[0, 1])
         pow_op = op.pow(2)
         assert isinstance(pow_op[0], hqml.SelectiveQubitRotation)
         assert pow_op[0].parameters == [1.0, 0.3]
 
     def test_simplify(self):
+        """Test the simplify method."""
         op = hqml.SelectiveQubitRotation(0, 0.3, 1, wires=[0, 1])
         simplified_op = op.simplify()
         assert isinstance(simplified_op, qml.Identity)
@@ -92,6 +78,7 @@ class TestSelectiveQubitRotation:
 @pytest.mark.unit
 class TestJaynesCummings:
     def test_init(self):
+        """Test the __init__ method."""
         op = hqml.JaynesCummings(0.5, 0.3, wires=[0, 1])
         assert op.name == "JaynesCummings"
         assert op.num_params == 2
@@ -100,18 +87,21 @@ class TestJaynesCummings:
         assert op.wires == qml.wires.Wires([0, 1])
 
     def test_adjoint(self):
+        """Test the adjoint method."""
         op = hqml.JaynesCummings(0.5, 0.3, wires=[0, 1])
         adj_op = op.adjoint()
         assert isinstance(adj_op, hqml.JaynesCummings)
         assert adj_op.parameters == [-0.5, 0.3]
 
     def test_pow(self):
+        """Test the pow method."""
         op = hqml.JaynesCummings(0.5, 0.3, wires=[0, 1])
         pow_op = op.pow(2)
         assert isinstance(pow_op[0], hqml.JaynesCummings)
         assert pow_op[0].parameters == [1.0, 0.3]
 
     def test_simplify(self):
+        """Test the simplify method."""
         op = hqml.JaynesCummings(0, 0.3, wires=[0, 1])
         simplified_op = op.simplify()
         assert isinstance(simplified_op, qml.Identity)
@@ -120,6 +110,7 @@ class TestJaynesCummings:
 @pytest.mark.unit
 class TestAntiJaynesCummings:
     def test_init(self):
+        """Test the __init__ method."""
         op = hqml.AntiJaynesCummings(0.5, 0.3, wires=[0, 1])
         assert op.name == "AntiJaynesCummings"
         assert op.num_params == 2
@@ -128,18 +119,21 @@ class TestAntiJaynesCummings:
         assert op.wires == qml.wires.Wires([0, 1])
 
     def test_adjoint(self):
+        """Test the adjoint method."""
         op = hqml.AntiJaynesCummings(0.5, 0.3, wires=[0, 1])
         adj_op = op.adjoint()
         assert isinstance(adj_op, hqml.AntiJaynesCummings)
         assert adj_op.parameters == [-0.5, 0.3]
 
     def test_pow(self):
+        """Test the pow method."""
         op = hqml.AntiJaynesCummings(0.5, 0.3, wires=[0, 1])
         pow_op = op.pow(2)
         assert isinstance(pow_op[0], hqml.AntiJaynesCummings)
         assert pow_op[0].parameters == [1.0, 0.3]
 
     def test_simplify(self):
+        """Test the simplify method."""
         op = hqml.AntiJaynesCummings(0, 0.3, wires=[0, 1])
         simplified_op = op.simplify()
         assert isinstance(simplified_op, qml.Identity)
@@ -148,6 +142,7 @@ class TestAntiJaynesCummings:
 @pytest.mark.unit
 class TestRabi:
     def test_init(self):
+        """Test the __init__ method."""
         op = hqml.Rabi(0.5, 0.3, wires=[0, 1])
         assert op.name == "Rabi"
         assert op.num_params == 2
@@ -156,18 +151,21 @@ class TestRabi:
         assert op.wires == qml.wires.Wires([0, 1])
 
     def test_adjoint(self):
+        """Test the adjoint method."""
         op = hqml.Rabi(0.5, 0.3, wires=[0, 1])
         adj_op = op.adjoint()
         assert isinstance(adj_op, hqml.Rabi)
         assert adj_op.parameters == [-0.5, 0.3]
 
     def test_pow(self):
+        """Test the pow method."""
         op = hqml.Rabi(0.5, 0.3, wires=[0, 1])
         pow_op = op.pow(2)
         assert isinstance(pow_op[0], hqml.Rabi)
         assert pow_op[0].parameters == [1.0, 0.3]
 
     def test_simplify(self):
+        """Test the simplify method."""
         op = hqml.Rabi(0, 0.3, wires=[0, 1])
         simplified_op = op.simplify()
         assert isinstance(simplified_op, qml.Identity)
@@ -176,6 +174,7 @@ class TestRabi:
 @pytest.mark.unit
 class TestConditionalDisplacement:
     def test_init(self):
+        """Test the __init__ method."""
         op = hqml.ConditionalDisplacement(0.5, 0.3, wires=[0, 1])
         assert op.name == "ConditionalDisplacement"
         assert op.num_params == 2
@@ -184,103 +183,22 @@ class TestConditionalDisplacement:
         assert op.wires == qml.wires.Wires([0, 1])
 
     def test_adjoint(self):
+        """Test the adjoint method."""
         op = hqml.ConditionalDisplacement(0.5, 0.3, wires=[0, 1])
         adj_op = op.adjoint()
         assert isinstance(adj_op, hqml.ConditionalDisplacement)
         assert adj_op.parameters == [-0.5, 0.3]
 
     def test_pow(self):
+        """Test the pow method."""
         op = hqml.ConditionalDisplacement(0.5, 0.3, wires=[0, 1])
         pow_op = op.pow(2)
         assert isinstance(pow_op[0], hqml.ConditionalDisplacement)
         assert pow_op[0].parameters == [1.0, 0.3]
 
     def test_simplify(self):
+        """Test the simplify method."""
         op = hqml.ConditionalDisplacement(0, 0.3, wires=[0, 1])
-        simplified_op = op.simplify()
-        assert isinstance(simplified_op, qml.Identity)
-
-
-@pytest.mark.unit
-class TestConditionalBeamsplitter:
-    def test_init(self):
-        op = hqml.ConditionalBeamsplitter(0.5, 0.3, wires=[0, 1, 2])
-        assert op.name == "ConditionalBeamsplitter"
-        assert op.num_params == 2
-        assert op.num_wires == 3
-        assert op.parameters == [0.5, 0.3]
-        assert op.wires == qml.wires.Wires([0, 1, 2])
-
-    def test_adjoint(self):
-        op = hqml.ConditionalBeamsplitter(0.5, 0.3, wires=[0, 1, 2])
-        adj_op = op.adjoint()
-        assert isinstance(adj_op, hqml.ConditionalBeamsplitter)
-        assert adj_op.parameters == [-0.5, 0.3]
-
-    def test_pow(self):
-        op = hqml.ConditionalBeamsplitter(0.5, 0.3, wires=[0, 1, 2])
-        pow_op = op.pow(2)
-        assert isinstance(pow_op[0], hqml.ConditionalBeamsplitter)
-        assert pow_op[0].parameters == [1.0, 0.3]
-
-    def test_simplify(self):
-        op = hqml.ConditionalBeamsplitter(0, 0.3, wires=[0, 1, 2])
-        simplified_op = op.simplify()
-        assert isinstance(simplified_op, qml.Identity)
-
-
-@pytest.mark.unit
-class TestConditionalTwoModeSqueezing:
-    def test_init(self):
-        op = hqml.ConditionalTwoModeSqueezing(0.5, 0.3, wires=[0, 1, 2])
-        assert op.name == "ConditionalTwoModeSqueezing"
-        assert op.num_params == 2
-        assert op.num_wires == 3
-        assert op.parameters == [0.5, 0.3]
-        assert op.wires == qml.wires.Wires([0, 1, 2])
-
-    def test_adjoint(self):
-        op = hqml.ConditionalTwoModeSqueezing(0.5, 0.3, wires=[0, 1, 2])
-        adj_op = op.adjoint()
-        assert isinstance(adj_op[0], hqml.ConditionalTwoModeSqueezing)
-        assert adj_op[0].parameters == [-0.5, 0.3]
-
-    def test_pow(self):
-        op = hqml.ConditionalTwoModeSqueezing(0.5, 0.3, wires=[0, 1, 2])
-        pow_op = op.pow(2)
-        assert isinstance(pow_op[0], hqml.ConditionalTwoModeSqueezing)
-        assert pow_op[0].parameters == [1.0, 0.3]
-
-    def test_simplify(self):
-        op = hqml.ConditionalTwoModeSqueezing(0, 0.3, wires=[0, 1, 2])
-        simplified_op = op.simplify()
-        assert isinstance(simplified_op, qml.Identity)
-
-
-@pytest.mark.unit
-class TestConditionalTwoModeSum:
-    def test_init(self):
-        op = hqml.ConditionalTwoModeSum(0.5, wires=[0, 1, 2])
-        assert op.name == "ConditionalTwoModeSum"
-        assert op.num_params == 1
-        assert op.num_wires == 3
-        assert op.parameters == [0.5]
-        assert op.wires == qml.wires.Wires([0, 1, 2])
-
-    def test_adjoint(self):
-        op = hqml.ConditionalTwoModeSum(0.5, wires=[0, 1, 2])
-        adj_op = op.adjoint()
-        assert isinstance(adj_op, hqml.ConditionalTwoModeSum)
-        assert adj_op.parameters == [-0.5]
-
-    def test_pow(self):
-        op = hqml.ConditionalTwoModeSum(0.5, wires=[0, 1, 2])
-        pow_op = op.pow(2)
-        assert isinstance(pow_op[0], hqml.ConditionalTwoModeSum)
-        assert pow_op[0].parameters == [1.0]
-
-    def test_simplify(self):
-        op = hqml.ConditionalTwoModeSum(0, wires=[0, 1, 2])
         simplified_op = op.simplify()
         assert isinstance(simplified_op, qml.Identity)
 
@@ -288,6 +206,7 @@ class TestConditionalTwoModeSum:
 @pytest.mark.unit
 class TestEchoedConditionalDisplacement:
     def test_init(self):
+        """Test the __init__ method."""
         op = hqml.EchoedConditionalDisplacement(0.5, 0, wires=[0, 1])
         assert op.name == "EchoedConditionalDisplacement"
         assert op.num_params == 2
@@ -296,16 +215,19 @@ class TestEchoedConditionalDisplacement:
         assert op.wires == qml.wires.Wires([0, 1])
 
     def test_adjoint(self):
+        """Test the adjoint method."""
         op = hqml.EchoedConditionalDisplacement(0.5, 0, wires=[0, 1])
         adj_op = op.adjoint()[0]
         assert adj_op.parameters == [-0.5, 0]
 
     def test_pow(self):
+        """Test the pow method."""
         op = hqml.EchoedConditionalDisplacement(0.5, 0.123, wires=[0, 1])
         pow_op = op.pow(4)[0]
         assert pow_op.parameters == [2, 0.123]
 
     def test_simplify(self):
+        """Test the simplify method."""
         op = hqml.EchoedConditionalDisplacement(0, 0.123, wires=[0, 1])
         simplified_op = op.simplify()
         assert isinstance(simplified_op, qml.Identity)
